requirements.txt
asgiref==3.7.1
celery==5.2.7
dj-database-url==2.0.0
Django==4.2.1
djangorestframework==3.14.0
django-filter==23.2
django-webpack-loader==2.0.0
psycopg2-binary==2.9.6
PyGithub==1.58.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
requests==2.31.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
whitenoise==6.4.0
